[#21] Extract chemistry source terms into module.
